Finish renaming validator
