+++ src/app/modules/organization/organization-users/create-user-dialog/create-user-dialog.component.ts
@@ -19,7 +19,7 @@ import { ONLY_DIGITS_AND_WHITESPACE_REGEX } from 'src/app/shared/constants/regex
 import { removeWhitespace } from 'src/app/shared/helpers/string.helpers';
 import { StartPreferenceChoice } from 'src/app/shared/models/organization/organization-user/start-preference.model';
 import { UserService } from 'src/app/shared/services/user.service';
-import { notDirtyAndEmptyStringValidator } from 'src/app/shared/validators/phone-number-length.validator';
+import { notDirtyAndEmptyStringValidator } from 'src/app/shared/validators/not-dirty-and-empty-string-validator';
 import { requiredIfDirtyValidator } from 'src/app/shared/validators/required-if-dirty.validator';
 import { OrganizationUserActions } from 'src/app/store/organization/organization-user/actions';
 import { selectOrganizationUserIsCreateLoading } from 'src/app/store/organization/organization-user/selectors';
