+++ src/app/modules/profile/profile.component.ts
@@ -1,5 +1,6 @@
+import { AsyncPipe, NgIf } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { Actions, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
 import { BehaviorSubject, combineLatestWith, map } from 'rxjs';
@@ -12,7 +13,7 @@ import {
 import { ValidatedValueChange } from 'src/app/shared/models/validated-value-change.model';
 import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
 import { UserService } from 'src/app/shared/services/user.service';
-import { phoneNumberLengthValidator } from 'src/app/shared/validators/phone-number-length.validator';
+import { notDirtyAndEmptyStringValidator } from 'src/app/shared/validators/not-dirty-and-empty-string-validator';
 import { OrganizationUserActions } from 'src/app/store/organization/organization-user/actions';
 import { UserActions } from 'src/app/store/user-store/actions';
 import {
@@ -20,16 +21,15 @@ import {
   selectUserDefaultUnit,
   selectUserOrganizationRights,
 } from 'src/app/store/user-store/selectors';
-import { ProfileComponentStore } from './profile.component-store';
 import { CardComponent } from '../../shared/components/card/card.component';
-import { NgIf, AsyncPipe } from '@angular/common';
-import { FormGridComponent } from '../../shared/components/form-grid/form-grid.component';
-import { OrgUnitSelectComponent } from '../../shared/components/org-unit-select/org-unit-select.component';
-import { TextBoxComponent } from '../../shared/components/textbox/textbox.component';
-import { TextBoxInfoComponent } from '../../shared/components/textbox-info/textbox-info.component';
-import { ParagraphComponent } from '../../shared/components/paragraph/paragraph.component';
 import { DropdownComponent } from '../../shared/components/dropdowns/dropdown/dropdown.component';
+import { FormGridComponent } from '../../shared/components/form-grid/form-grid.component';
 import { LoadingComponent } from '../../shared/components/loading/loading.component';
+import { OrgUnitSelectComponent } from '../../shared/components/org-unit-select/org-unit-select.component';
+import { ParagraphComponent } from '../../shared/components/paragraph/paragraph.component';
+import { TextBoxInfoComponent } from '../../shared/components/textbox-info/textbox-info.component';
+import { TextBoxComponent } from '../../shared/components/textbox/textbox.component';
+import { ProfileComponentStore } from './profile.component-store';
 
 @Component({
   templateUrl: 'profile.component.html',
@@ -59,7 +59,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
     map(([organizationRights, organizationUuid]) => {
       if (!organizationRights) return false;
       return organizationRights.some((right) => right.organizationUuid === organizationUuid);
-    }),
+    })
   );
   public readonly userDefaultUnit$ = this.store.select(selectUserDefaultUnit);
 
@@ -72,7 +72,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
     firstName: new FormControl<string | undefined>(undefined, Validators.required),
     lastName: new FormControl<string | undefined>(undefined, Validators.required),
     email: new FormControl<string | undefined>(undefined, [Validators.required, Validators.email]),
-    phoneNumber: new FormControl<string | undefined>(undefined, phoneNumberLengthValidator()),
+    phoneNumber: new FormControl<string | undefined>(undefined, notDirtyAndEmptyStringValidator()),
     defaultStartPreference: new FormControl<StartPreferenceChoice | undefined>(undefined),
     defaultOrganizationUnit: new FormControl<APIIdentityNamePairResponseDTO | undefined>(undefined),
   });
@@ -83,7 +83,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
     private store: Store,
     private componentStore: ProfileComponentStore,
     private actions$: Actions,
-    private userService: UserService,
+    private userService: UserService
   ) {
     super();
   }
@@ -121,7 +121,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
           this.editForm.patchValue({
             defaultOrganizationUnit: defaultUnitToPatch,
           });
-        }),
+        })
     );
 
     this.subscriptions.add(
@@ -129,7 +129,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
         this.currentEmail = email;
         this.alreadyExists$.next(false);
         this.onChange({ email });
-      }),
+      })
     );
 
     this.subscriptions.add(
@@ -140,7 +140,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
         }
 
         this.changeEmailValidityState(true);
-      }),
+      })
     );
 
     this.subscriptions.add(
@@ -150,7 +150,7 @@ export class ProfileComponent extends BaseComponent implements OnInit {
           if (organizationUnit.uuid === currentUnitUuid) return;
 
           this.currentDefaultUnitUuid$.next(organizationUnit.uuid);
-        }),
+        })
     );
   }
